Initial Commit - Cmake

+++ source/Console.cpp
@@ -73,9 +73,10 @@ void Console::draw() {
     else 
         DrawRectangleRounded({m_position.x, m_position.y, m_size.x, m_size.y}, m_roundness, m_segment, m_normal_color);
     BeginScissorMode(m_position.x, m_position.x, m_size.x, m_size.y);
-        if (m_line_choice>=0) 
-            DrawRectangle(m_position.x, m_position.y+m_font_size*m_line_choice+m_delta.y, m_size.x, m_font_size, m_color_line);
-        for (int i = 0; i<m_list.size(); i++) 
-            DrawTextEx(m_font, m_list[i].c_str(), m_origin+m_delta+m_position+(Vector2){0, m_font_size*i}, m_font_size, m_spacing, m_text_color);
+        if (m_line_choice >= 0) 
+            DrawRectangle(m_position.x, m_position.y + m_font_size * m_line_choice + m_delta.y, m_size.x, m_font_size, m_color_line);
+        for (size_t i = 0; i < m_list.size(); i++) 
+            DrawTextEx(m_font, m_list[i].c_str(),
+                       m_origin + m_delta + m_position + Vector2{0, m_font_size * (float)i}, m_font_size, m_spacing, m_text_color);
     EndScissorMode();
 }
